backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
google-genai>=1.0
lightgbm>=4.0
shap>=0.43.0
PyYAML>=6.0
SQLAlchemy>=2.0
alembic>=1.13
redis>=5.0
rq>=1.16,<2
python-jose[cryptography]>=3.3.0
Authlib>=1.3
psycopg[binary]>=3.1
prometheus-client>=0.20
opentelemetry-api>=1.24
opentelemetry-sdk>=1.24
opentelemetry-exporter-otlp>=1.24
opentelemetry-instrumentation-fastapi>=0.45b0
